Updating the _get_authors() function
While searching for the 'author' tag in the XML file, the findall function was not being able to find it because the 'author' tag isn't a direct child of the 'Record' tag as part of the Tree structure of the XML file. Hence, the findall function returns an empty list for the variable 'authors_xml' which leads to an attribute error while trying to get the first and last names of authors. The proposed change modifies the input of the findall function to ensure that it searches the relevant structure of the XML tree for the 'author' tag.